Code refactoring
- move Collect() logic to collect() func
- move collect() to the collect.go
- create a file per query
  - collect_server_status.go
  - collect_dbstats.go
  - collect_replsetstatus.go
