+++ modules/mongodb/charts.go
@@ -530,86 +530,3 @@ var (
 		Ctx:   "mongodb." + replicationNodePing,
 	}
 )
-
-// updateDBStatsCharts adds dimensions for new databases and
-// removes for dropped
-func (m *Mongo) updateDBStatsCharts(databases []string) {
-	// remove dims for not existing databases
-	m.removeDatabasesFromDBStatsCharts(databases)
-
-	// add dimensions for new databases
-	for _, database := range sliceDiff(databases, m.discoveredDBs) {
-		for _, chart := range *m.chartsDbStats {
-			id := chart.ID + "_" + database
-			err := chart.AddDim(&module.Dim{ID: id, Name: database, Algo: module.Absolute})
-			if err != nil {
-				m.Warningf("failed to add dim: %s, %v", id, err)
-				continue
-			}
-			chart.MarkNotCreated()
-		}
-	}
-
-	// update the cache
-	m.discoveredDBs = databases
-}
-
-// removeDatabasesFromDBStatsCharts removes dimensions from dbstats
-// charts for dropped databases
-func (m *Mongo) removeDatabasesFromDBStatsCharts(newDatabases []string) {
-	diff := sliceDiff(m.discoveredDBs, newDatabases)
-	for _, name := range diff {
-		for _, chart := range *m.chartsDbStats {
-			id := chart.ID + "_" + name
-			err := chart.MarkDimRemove(id, true)
-			if err != nil {
-				m.Warningf("failed to remove dimension %s with error: %v", id, err)
-				continue
-			}
-			chart.MarkNotCreated()
-		}
-	}
-}
-
-// removeReplicaSetMember removes dimensions for not existing
-// replica set members
-func (m *Mongo) removeReplicaSetMembers(newMembers []string) {
-	diff := sliceDiff(m.replSetMembers, newMembers)
-	for _, name := range diff {
-		for _, v := range []struct{ chartID, dimPrefix string }{
-			{replicationLag, replicationLagDimPrefix},
-			{replicationHeartbeatLatency, replicationHeartbeatLatencyDimPrefix},
-			{replicationNodePing, replicationNodePingDimPrefix},
-		} {
-			id := v.dimPrefix + name
-			if !m.replSetDimsEnabled[id] {
-				continue
-			}
-			delete(m.replSetDimsEnabled, id)
-
-			chart := m.charts.Get(v.chartID)
-			if chart != nil {
-				if err := chart.MarkDimRemove(id, true); err != nil {
-					m.Warningf("failed to remove dimension: %v", err)
-				}
-			} else {
-				m.Warningf("failed to remove dimension:%s. job doesn't have chart: %s", id, v.chartID)
-			}
-		}
-	}
-}
-
-// sliceDiff calculates the diff between to slices
-func sliceDiff(slice1, slice2 []string) []string {
-	mb := make(map[string]struct{}, len(slice2))
-	for _, x := range slice2 {
-		mb[x] = struct{}{}
-	}
-	var diff []string
-	for _, x := range slice1 {
-		if _, found := mb[x]; !found {
-			diff = append(diff, x)
-		}
-	}
-	return diff
-}
